refactor: 모바일 정보수정 기능 연결을 위한 관련로직,타입 분리
