+++ src/hooks/useUpdateProfile.ts
@@ -1,15 +1,32 @@
 import { useCallback, useState } from 'react';
-import { useSetRecoilState } from 'recoil';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useRecoilState, useSetRecoilState } from 'recoil';
 import { mypageInfoButtonActiveState, userInfoState } from '../recoil/atoms';
 import { contactValidation, nicknameValidation } from 'utils/validation';
+import { updateUserInfoData } from 'apis/mypageUsers';
+import useAuth from './useAuth';
+import useToastPopup from './useToastPopup';
 
+// 마이페이지 개인정보를 위한 훅
 const useUpdateProfile = () => {
-  const setUserInfo = useSetRecoilState<UserInfo>(userInfoState);
+  const [userInfo, setUserInfo] = useRecoilState<UserInfo>(userInfoState);
   const setActiveInfoBtn = useSetRecoilState(mypageInfoButtonActiveState);
-
+  const { uid } = useAuth();
   const [validationMessage, setValidationMessage] = useState<string>('');
+  const { showToast, ToastMessage, handleToastPopup } = useToastPopup();
   const [contactValidationMessage, setContactValidationMessage] =
     useState<string>('');
+
+  // 정보 수정 파이어베이스 mutatation
+  const queryClient = useQueryClient();
+  const { mutate: updateUserInfoMutate } = useMutation(
+    () => updateUserInfoData(uid, userInfo),
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries(['users', uid]);
+      },
+    },
+  );
 
   // 텍스트 인풋 변경 핸들러
   const handleInputChange = useCallback(
@@ -44,10 +61,35 @@ const useUpdateProfile = () => {
     [],
   );
 
+  // 정보완료 버튼 유효성 검사
+  const checkInfoValidation = () => {
+    const nickname = userInfo.displayName;
+    if (nickname.length < 2 || nickname.length > 7) {
+      handleToastPopup('닉네임은 2자 이상 7자 이하로 입력해주세요.');
+      return false;
+    }
+
+    if (userInfo.email && !contactValidation(userInfo.email)) {
+      handleToastPopup('연락처를 올바르게 입력해주세요.');
+      return false;
+    }
+
+    if (userInfo.positions.length === 0) {
+      handleToastPopup('포지션을 선택해주세요.');
+      return false;
+    }
+    return true;
+  };
+
   return {
     validationMessage,
     handleInputChange,
     contactValidationMessage,
+    updateUserInfoMutate,
+    showToast,
+    ToastMessage,
+    handleToastPopup,
+    checkInfoValidation,
   };
 };
